Phase 1: Extract document form hook and LineItemsTable; refactor Create/Edit to use them
